move _keep_coeff to mul since it is a special Mul routine
