+++ sympy/core/add.py
@@ -687,33 +687,7 @@ class Add(AssocOp):
         """
         return Add(*[_keep_coeff(*a.as_content_primitive()) for a in self.args]).primitive()
 
-def _keep_coeff(coeff, factors):
-    """Return ``coeff*factors`` unevaluated if necessary."""
-
-    if not coeff.is_Number:
-        if factors.is_Number:
-            factors, coeff = coeff, factors
-        else:
-            return coeff*factors
-    if coeff == 1:
-        return factors
-    elif coeff == -1: # don't keep sign?
-        return -factors
-    elif factors.is_Add:
-        return Mul._from_args((coeff, factors))
-    elif factors.is_Mul:
-        margs = list(factors.args)
-        if margs[0].is_Number:
-            margs[0] *= coeff
-            if margs[0] == 1:
-                margs.pop(0)
-        else:
-            margs.insert(0, coeff)
-        return Mul._from_args(margs)
-    else:
-        return coeff*factors
-
 from function import FunctionClass
-from mul import Mul
+from mul import Mul, _keep_coeff
 from power import Pow
 from sympy.core.numbers import Rational
